bug fix: pruned variant ids in pca correlation were incorrect
The GDS file resulting from the LD pruning workflow has variant
ids numbered 1:n that do not match the original variant ids. Thus
it does not work to take the ids from the pruned file, but they
must be passed as an additional argument. Since the rdata files
output by the workflow are per chromosome, take an array of
ids and scatter on both full gds files and variant include files.